Rename sds calls so they don't conflict in Redis.
Hiredis and Redis both contain the sds library which can cause hard to
debug issues when there are ABI mismatches between the two copies.

This recently happened when merging v1.0.0 into Redis because Hiredis
updated sdsrange to return a value.  When called from within Redis, this
abi mismatch caused random failures.

It's not the prettiest fix, but this commit simply renames the sds
types/functions in hiredis so they can't conflict with those built in
redis.